+++ Modules/Intent.Modules.Common.CSharp/release-notes.md
@@ -2,6 +2,7 @@
 
 - Improvement: Added support for Razor file code management.
 - Fixed: Nullable aware assignments of instantiated objects via mapping systems not resolving replacements correctly, leading to uncompilable code.
+- Fixed: Mapping system unable to find constructor when multiple partial C# files are in play.
 
 ### Version 3.7.6